Actualizando la logica de inicio del juego
